requirements.txt
streamlit==1.28.1
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
openai==1.3.7
faiss-cpu==1.7.4
pypdf==3.17.1
python-dotenv==1.0.0
tiktoken==0.5.1
numpy==1.24.3
pandas==2.0.3 